requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
PyYAML==5.1.1
requests==2.22.0
slackclient==1.3.2
urllib3==1.25.3
websocket-client==0.54.0
